* feature/mission_system:
  Mission: Development - Current value is now a float in order to have more precision when tracking certain objectives (i.e. time). Mission Kill Objective: Development - Full implementation. Mission Survive Objective: Development - Full implementation. Mission Defintions: Content - Created a sample of each mission type.
  Mission Manager: Cleanup - Removed some debug logs.


Former-commit-id: ab6a5c8296d25c80e515eb55c1534f23022c29a0

+++ Assets/Game/Model/Missions/Mission.cs
@@ -27,7 +27,7 @@ public class Mission {
 	[Serializable]
 	public class SaveData {
 		public string sku = "";		// Mission def's sku
-		public int currentValue = 0;	// Objective's current value - only relevant for long-term missions, but save it always anyway
+		public float currentValue = 0f;	// Objective's current value - only relevant for long-term missions, but save it always anyway
 	}
 
 	//------------------------------------------------------------------//
